Assignment 7
Large Display Changes..

I got rid of the transparent background behind the products I also
changed how it was two columns in all the products because on small
display I wanted it to just be center aligned

Was not responding well and looked bad.

Small Display Changes..

Changed the way footer looks with the back to top stacked on top of the
social media icons

It looked more pleasing to the eye.

I changed the way the map looked I didn’t want it to expand the full
with

Looked better with it aligning to the contact page.

I changed the quote to take up more space

It was so tiny and hard to see.